Improve 404 errors
If we let them flow through our custom exception handling normally
instead of rescueing and rising custom exceptions, we get the same
behavior in production (a 404 page), but better error pages in
development, with code context & backtrace information, instead of a
single "Not found" message.

And simpler code, of course.