Extend controller only when needed
In the case of the participatory process index page, we are still
extending the controller with the `NeedsParticipatoryProcess` module.
This has several drawbacks:

* It includes a lot of methods that don't make sense for the index page,
like `current_participatory_process`.
* It requires hacks to "undo" the inclusion like the `skip_after_action`
hook in the controller.
* It makes the index action touch the DB unnecessarily. When authorizing
the action, the `ability_context` tries to call
`current_participatory_process`, and since the method is defined, DB is
touched. Things work "by chance" because the query returns `nil`, but
it's incorrect and inefficient.

This approach should fix these problems.