Fail early on missing participatory process groups
Otherwise we fail when rendering the template. This is less efficient
since we do more unnecessary work, and it raises a different error
(`ActionView::Template::Error`) than the other "not found" cases. I'm
guessing Rails treats these just like `RoutingError` or `RecordNotFound`
error, but it feels better to be consistent.